tests/wpt/tests/html/canvas/element/text/2d.text.drawing.style.wordSpacing.measure.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.text.drawing.style.wordSpacing.measure</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.text.drawing.style.wordSpacing.measure</h1>
<p class="desc">Testing word spacing with different length units</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Testing word spacing with different length units");
_addTest(function(canvas, ctx) {

  _assertSame(ctx.letterSpacing, '0px', "ctx.letterSpacing", "'0px'");
  _assertSame(ctx.wordSpacing, '0px', "ctx.wordSpacing", "'0px'");
  var width_normal = ctx.measureText('Hello World, again').width;

  function test_word_spacing(value, difference_spacing, epsilon) {
    ctx.wordSpacing = value;
    _assertSame(ctx.letterSpacing, '0px', "ctx.letterSpacing", "'0px'");
    _assertSame(ctx.wordSpacing, value, "ctx.wordSpacing", "value");
    width_with_word_spacing = ctx.measureText('Hello World, again').width;
    assert_approx_equals(width_with_word_spacing, width_normal + difference_spacing, epsilon, "word spacing doesn't work.");
  }

  // The first value is the word Spacing to be set, the second value the
  // change in length of string 'Hello World', note that there are 2 words
  // in 'Hello World, again', so the length difference is always wordSpacing * 2.
  // and the third value is the acceptable differencee for the length change,
  // note that unit such as 1cm/1mm doesn't map to an exact pixel value.
  test_cases = [['3px', 6, 0.1],
               ['5px', 10, 0.1],
               ['-2px', -4, 0.1],
               ['1em', 20, 0.1],
               ['-0.5em', -10, 0.1],
               ['1in', 192, 0.1],
               ['-0.1cm', -7.57, 0.2],
               ['-0.6mm', -4.54, 0.2]]

  for (const test_case of test_cases) {
    test_word_spacing(test_case[0], test_case[1], test_case[2]);
  }

});
</script>
