third_party/blink/web_tests/platform/win/external/wpt/html/canvas/element/text/2d.text.drawing.style.wordSpacing.measure-expected.txt

This is a testharness.js-based test.
FAIL Testing word spacing with different length units assert_approx_equals: word spacing doesn't work. expected 97.77777777777777 +/- 0.1 but got 88
Harness: the test ran to completion.
